joy_divison_maker_backend/requirements.txt
asgiref==3.8.1
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
numpy==2.1.0
pillow==10.4.0
sqlparse==0.5.1
